Added Tempory Channel fetures and Dynamic Help command
Adds Dynmaic Help command where the general purpose is that it autolmatically adds the command in the help list without having the change the code of help command file.

Adds Tempory channel feutures where the general purpose is to create tempory voice and text channel for the user so they can have a private group togeather on all servers.